chore(core): move `ActorStatus` and related code into the separate file (`actor_status.rs`), fix comments and style
